[ISSUE] #32 Renamed controllers-package to controller
