+++ src/main/kotlin/com/schulverwaltung/controller/CSVImportHistoryController.kt
@@ -63,6 +63,4 @@ class CSVImportHistoryController(
             return@newTransactionScope result
         }
     }
-
-
 }
