+++ src/main/kotlin/com/schulverwaltung/controller/interfaces/ISecretController.kt
