+++ src/main/kotlin/com/schulverwaltung/controller/interfaces/ISecretaryController.kt
