+++ src/main/kotlin/com/schulverwaltung/controller/interfaces/IStudentController.kt
